test: move jenkins mocks
