+++ test/test_jenkins.py
@@ -12,7 +12,7 @@ from bob.errors import BuildError
 from bob.state import BobState
 from unittest.mock import patch
 
-from jenkins.jenkins_mock import JenkinsMock
+from mocks.jenkins_mock import JenkinsMock
 
 from bob.cmds.jenkins import doJenkins
 from bob.state import finalize
